# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ECOnexão API
  description: API de domínio do aplicativo ECOnexão para turismo sustentável e rotas
    ecológicas.
  version: 1.0.0
paths:
...
  /api/v1/admin/workflow/{resource_type}/{resource_id}/transition:
    post:
      tags:
      - Admin Workflow
      summary: Transicionar estado editorial
      description: Transicionar o estado de um recurso na máquina de estados editorial
        com validação de publish guard e concorrência.
      operationId: transition_resource_status_api_v1_admin_workflow__resource_type___resource_id__transition_post
      security:
      - HTTPBearer: []
      parameters:
      - name: resource_type
        in: path
        required: true
        schema:
          type: string
          title: Resource Type
      - name: resource_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Resource Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusTransitionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusTransitionEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/workflow/{resource_type}/{resource_id}/publish-guard:
    get:
      tags:
      - Admin Workflow
      summary: Consultar Publish Guard
      description: Consultar os critérios e elegibilidade do Publish Guard para um
        recurso.
      operationId: get_publish_guard_status_api_v1_admin_workflow__resource_type___resource_id__publish_guard_get
      security:
      - HTTPBearer: []
      parameters:
      - name: resource_type
        in: path
        required: true
        schema:
          type: string
          title: Resource Type
      - name: resource_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Resource Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishGuardResultEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/alerts:
    get:
      tags:
      - Admin Workflow
      summary: Listar alertas editoriais
      description: Listar alertas editoriais ativos ou resolvidos com suporte a filtros
        de rota e severidade.
      operationId: list_editorial_alerts_api_v1_admin_alerts_get
      security:
      - HTTPBearer: []
      parameters:
      - name: route_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Filtrar por rota
          title: Route Id
        description: Filtrar por rota
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filtrar por severidade
          title: Severity
        description: Filtrar por severidade
      - name: is_active
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Filtrar por estado ativo
          title: Is Active
        description: Filtrar por estado ativo
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditorialAlertListEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Admin Workflow
      summary: Criar alerta editorial
      operationId: create_editorial_alert_api_v1_admin_alerts_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditorialAlertCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditorialAlertEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/alerts/{alert_id}/resolve:
    post:
      tags:
      - Admin Workflow
      summary: Resolver alerta editorial
      description: Resolver um alerta editorial ativo fornecendo uma nota explicativa
        de resolução.
      operationId: resolve_editorial_alert_api_v1_admin_alerts__alert_id__resolve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlertResolveRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditorialAlertEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/reconciliation/candidates:
    get:
      tags:
      - Admin Workflow
      summary: Listar candidatos de reconciliação
      description: Listar candidatos a duplicata para revisão humana e reconciliação
        editorial.
      operationId: list_reconciliation_candidates_api_v1_admin_reconciliation_candidates_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationCandidateListEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/reconciliation/{candidate_id}/decision:
    post:
      tags:
      - Admin Workflow
      summary: Decidir reconciliação
      description: Registrar decisão editorial (accept, reject, merge) sobre candidato
        de reconciliação com justificativa auditada.
      operationId: decide_reconciliation_candidate_api_v1_admin_reconciliation__candidate_id__decision_post
      security:
      - HTTPBearer: []
      parameters:
      - name: candidate_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Candidate Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconciliationDecisionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationDecisionEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/reconciliation/{candidate_id}/compensate:
    post:
      tags:
      - Admin Workflow
      summary: Compensar merge de reconciliação
      operationId: compensate_reconciliation_merge_api_v1_admin_reconciliation__candidate_id__compensate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: candidate_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Candidate Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconciliationCompensationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationDecisionEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/alerts/{alert_id}:
    put:
      tags:
      - Admin Workflow
      summary: Atualizar alerta editorial
      operationId: update_editorial_alert_api_v1_admin_alerts__alert_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Alert Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditorialAlertUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditorialAlertEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AlertResolveRequest:
      properties:
        resolution_note:
          type: string
          minLength: 3
          title: Resolution Note
          description: Nota explicativa da resolução do alerta
      type: object
      required:
      - resolution_note
      title: AlertResolveRequest
...
    EditorialAlertCreateRequest:
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        message:
          type: string
          minLength: 1
          title: Message
        severity:
          type: string
          enum:
          - info
          - warning
          - critical
          title: Severity
          default: info
        source:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Source
        starts_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Starts At
        ends_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ends At
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
        is_active:
          type: boolean
          title: Is Active
          default: true
        route_id:
          type: string
          format: uuid
          title: Route Id
      type: object
      required:
      - title
      - message
      - route_id
      title: EditorialAlertCreateRequest
    EditorialAlertEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/EditorialAlertSchema'
      type: object
      required:
      - data
      title: EditorialAlertEnvelope
    EditorialAlertListEnvelope:
      properties:
        data:
          items:
            $ref: '#/components/schemas/EditorialAlertSchema'
          type: array
          title: Data
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      type: object
      required:
      - data
      - meta
      title: EditorialAlertListEnvelope
...
    EditorialAlertUpdateRequest:
      properties:
        title:
          anyOf:
          - type: string
            maxLength: 255
            minLength: 1
          - type: 'null'
          title: Title
        message:
          anyOf:
          - type: string
            minLength: 1
          - type: 'null'
          title: Message
        severity:
          anyOf:
          - type: string
            enum:
            - info
            - warning
            - critical
          - type: 'null'
          title: Severity
        source:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Source
        starts_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Starts At
        ends_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Ends At
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
      type: object
      title: EditorialAlertUpdateRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginationMeta:
      properties:
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
      type: object
      required:
      - total
      - limit
      title: PaginationMeta
    PublishGuardResultEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/PublishGuardResultSchema'
      type: object
      required:
      - data
      title: PublishGuardResultEnvelope
    PublishGuardResultSchema:
      properties:
        resource_type:
          type: string
          title: Resource Type
        resource_id:
          type: string
          format: uuid
          title: Resource Id
        current_status:
          type: string
          title: Current Status
        is_eligible:
          type: boolean
          title: Is Eligible
        missing_requirements:
          items:
            type: string
          type: array
          title: Missing Requirements
        warnings:
          items:
            type: string
          type: array
          title: Warnings
      type: object
      required:
      - resource_type
      - resource_id
      - current_status
      - is_eligible
      title: PublishGuardResultSchema
    ReconciliationCandidateListEnvelope:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ReconciliationCandidateSchema'
          type: array
          title: Data
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      type: object
      required:
      - data
      - meta
      title: ReconciliationCandidateListEnvelope
...
    ReconciliationCompensationRequest:
      properties:
        reason:
          type: string
          minLength: 5
          title: Reason
          description: Justificativa da compensação
      type: object
      required:
      - reason
      title: ReconciliationCompensationRequest
    ReconciliationDecisionEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/ReconciliationDecisionSchema'
      type: object
      required:
      - data
      title: ReconciliationDecisionEnvelope
    ReconciliationDecisionRequest:
      properties:
        decision:
          type: string
          enum:
          - accept
          - reject
          - merge
          title: Decision
          description: Decisão editorial sobre o candidato
        reason:
          type: string
          minLength: 5
          title: Reason
          description: Justificativa da decisão editorial
        target_actor_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Target Actor Id
          description: UUID do ator primário no caso de mesclagem (merge)
      type: object
      required:
      - decision
      - reason
      title: ReconciliationDecisionRequest
...
    StatusTransitionEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/StatusTransitionSchema'
      type: object
      required:
      - data
      title: StatusTransitionEnvelope
    StatusTransitionRequest:
      properties:
        target_status:
          type: string
          enum:
          - draft
          - review
          - published
          - archived
          title: Target Status
          description: Novo estado editorial desejado
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
          description: Motivo ou justificativa da alteração de estado
        expected_version:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Expected Version
          description: Versão esperada para controle de concorrência otimista
      type: object
      required:
      - target_status
      title: StatusTransitionRequest
    StatusTransitionSchema:
      properties:
        resource_type:
          type: string
          title: Resource Type
        resource_id:
          type: string
          format: uuid
          title: Resource Id
        previous_status:
          type: string
          title: Previous Status
        new_status:
          type: string
          title: New Status
        version:
          type: integer
          title: Version
        audit_log_id:
          type: string
          format: uuid
          title: Audit Log Id
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - resource_type
      - resource_id
      - previous_status
      - new_status
      - version
      - audit_log_id
      - updated_at
      title: StatusTransitionSchema
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer